requirements.txt
requests
Flask==0.10.1
SQLAlchemy==1.0
gevent==1.1b3
Flask-SocketIO==0.6.0
paho-mqtt==1.1
Flask-SQLAlchemy
Flask-Moment==0.4.0
Flask-WTF==0.12
WTForms==2.0.2
gunicorn